+++ PSFramework/internal/tepp/scripts/callback-name.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'PSFramework.Callback.Name' -ScriptBlock {
+	(Get-PSFCallback).Name | Select-Object -Unique
+}